Add avi to mediaExts in options.js
This makes .avi -files visible for the file open dialog.

The theory is, that this gives better user experience.
Given that more often than not, only thing needed to make the "avi" files working is to rename the
extension to something "supported", like .mp4, it is reasonable to assume that the inverse would work as well
(i.e. don't refuse to open .avi files just because of their extension).